Move CuraStage back to where plugins expect it.
