+++ plugins/MonitorStage/MonitorStage.py
@@ -2,7 +2,7 @@
 # Cura is released under the terms of the LGPLv3 or higher.
 import os.path
 from UM.Application import Application
-from cura.UI.CuraStage import CuraStage
+from cura.Stages.CuraStage import CuraStage
 
 
 ##  Stage for monitoring a 3D printing while it's printing.
